smart-city-backend/requirements.txt
asgiref==3.11.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
djangorestframework==3.16.1
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-pubsub==2.33.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
importlib_metadata==8.7.0
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
proto-plus==1.26.1
protobuf==6.33.1
psycopg==3.3.0
psycopg-binary==3.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.4
typing_extensions==4.15.0
urllib3==2.5.0
zipp==3.23.0
